PDF Of Schematic
The schematic shows the solder points of the input and output, along with the formula used to determine the value of R1. I set the value of R2 to 10k, VFB is 0.8 according to the datasheet, and we want the output voltage to be between 5V and close to 0V. Using a 50k pot we can get the range down to 1.2V, which for low-voltage breadboard circuits is sufficient enough as a proof of concept. The switching frequency can have a continuous output current of 2A.